requirements.txt
# ===================================================================
# REQUIREMENTS.TXT - AFEPANOU MARKETPLACE BACKEND
# ===================================================================
# Compatible: Python 3.10+, Django 4.x, PostgreSQL, Redis, Railway
# Marketplace e-commerce haïtienne avec intégration MonCash + Celery
# ===================================================================

# ===================================================================
# DJANGO CORE FRAMEWORK
# ===================================================================
Django
djangorestframework
django-cors-headers

# ===================================================================
# DATABASE & ORM
# ===================================================================
psycopg2-binary
dj-database-url

# ===================================================================
# AUTHENTICATION & SECURITY
# ===================================================================
djangorestframework-simplejwt
cryptography

# ===================================================================
# CACHING & PERFORMANCE (REDIS)
# ===================================================================
redis
django-redis

# ===================================================================
# CELERY BACKGROUND TASKS (COMPATIBLE PYTHON 3.10+)
# ===================================================================
celery
django-celery-beat
django-celery-results
flower
kombu

# ===================================================================
# FILE STORAGE & MEDIA (BACKBLAZE B2)
# ===================================================================
django-storages
boto3
Pillow

# ===================================================================
# API & FILTERING
# ===================================================================
django-filter
drf-spectacular

# ===================================================================
# HTTP REQUESTS (MONCASH INTEGRATION)
# ===================================================================
requests

# ===================================================================
# ENVIRONMENT & CONFIGURATION
# ===================================================================
python-dotenv
python-decouple
django-environ

# ===================================================================
# STATIC FILES & DEPLOYMENT
# ===================================================================
whitenoise
gunicorn

# ===================================================================
# UTILITIES & HELPERS
# ===================================================================
python-slugify
python-dateutil
pytz
babel
py-moneyed

# ===================================================================
# ADMIN INTERFACE & CONTENT MANAGEMENT
# ===================================================================
django-extensions
django-colorfield
django-taggit

# ===================================================================
# SECURITY & RATE LIMITING
# ===================================================================
django-ratelimit
django-csp

# ===================================================================
# MONITORING & HEALTH CHECKS
# ===================================================================
django-health-check
psutil
setproctitle

# ===================================================================
# PRODUCTION OPTIMIZATIONS
# ===================================================================
django-compressor

# ===================================================================
# DEVELOPMENT & TESTING (OPTIONNEL)
# ===================================================================
pytest
pytest-django
coverage
ipython

# ===================================================================
# CODE QUALITY (OPTIONNEL)
# ===================================================================
flake8
black
isort

# ===================================================================
# ASYNC SUPPORT (OPTIONNEL POUR CELERY)
# ===================================================================
eventlet
gevent
